Move schema mount machinery into the pi-tui renderer
Relocate SchemaCallComponent/SchemaResultComponent and mountCall/mountResult
out of schema.ts into renderers/pi-tui/schema-mount.ts. schema.ts is now pure
render-schema data plus the renderer-agnostic ANSI projection
(segmentsToString/iconString/statusSuffix/renderBody, now exported for renderers).

mountCall/mountResult are re-exported from schema.ts to preserve the published
@guanyilun/ashi/render surface; hooks.ts imports them from the new location.
First step toward a swappable Renderer behind ashi — the schema contract no
longer depends on pi-tui.
